+++ src/ApolloState.m
@@ -1,6 +1,7 @@
 #import "ApolloState.h"
 
 NSString *sRedditClientId = nil;
+NSString *sRedditClientSecret = nil;
 NSString *sImgurClientId = nil;
 NSString *sRedirectURI = nil;
 NSString *sUserAgent = nil;
@@ -20,6 +21,7 @@ NSInteger sUnmuteCommentsVideos = 0; // 0=Default, 1=Remember from Full Screen, 
 BOOL sProxyImgurDDG = NO;
 BOOL sShowUserAvatars = NO;
 BOOL sUseProfileAvatarTabIcon = NO;
+BOOL sShowSubredditHeaders = NO;
 BOOL sAutoHideTabBarShowOnIdle = NO;
 BOOL sModernSubredditDividers = YES;
 BOOL sEnableInlineImages = NO;